Phase one of complete website overhaul
The old version was endearing and is tagged as v0.1 for my records,
but doesn't represent a good state for web design. I have learned
a great deal since then and am working on completely overhauling
both the core content and design of the website.

This is the first step (stripping out useless HTML and resetting the
style back to nothing), from which I can build a foundation.