Begin footer styling for increased content
Adjust back to top button to be an arrow and add styling to bring
closer to the rest of website design. Add grid layout to footer to prepare
for resources, social media, and contact fields. Separate copyright info
to bottom of footer so it's there but out of the way of real content.