Add template for home page, adjust styles
Create several fields on home page plus default spacing to begin
work on a nice, scrolling home page.

TODO: Decide on the importance (or not) of the about tab. Redo parts
of the design to ensure the stylesheet flows cleanly and semantically,
and also to ensure professionalism and cleanliness in design. It may be
too cluttered or busy currently.