Toggle hide on social media
Display button to display or hide the social media channels linked
at the top of each page. Reduce clutter at the top of the page and encourage
exploration of the remainder of the site rather than immediately hitting
you with social options.

Style social menu to have better spacing and a smooth transition. Note:
the current transition is a placeholder and will be updated.