Make table of contents collapsible
Add collapsible feature to table of contents to preserve scrolling space
by default but allow the reader to expand the toc if valuable to them.

Create animations to drop the table of content li into place in sequence,
but only for readers who allow animation (held in a media query).