Add read time calculation to blog posts
Use JavaScript to calculate the read time of a page based on average
words per minute read. Captures all the words in a post and translates
that into the estimated read time, then posts it at the top of the page
below the title.

Create basic styles for read time.