Refactor subscription-form
Move styles for subscription form at the bottom of each page to a class, .form-widget, instead of relying on #subscription-form for styles AND as a JavaScript hook. Instead, refactor to separate concerns, such that the JS targets an ID with no styles, and the CSS targets a class with all the styles but no JS hook.