Optimization & color refactor
Optimize images to be webp format and be a smaller size closer to how they appear in the final product of the website to release load times on older devices.

Reformat colors across both themes to be monochromatic with a different palette depending on light/dark mode, rather than using a mix of each.

Change subscription box to adjust colors with theme instead of using static colors.