Add projects subdirectory
Create new folder for projects, a landing page which will include the official pages for things such as Pokewrap and other published software I create. This may be the precursor to my portfolio, or it may just be a clever way for me to share what I'm doing and microblog about it.