Refine Projects page, add Pokewrap page
Change styles in Projects home page to be more visually appealing and complete. Add a button linking to the Pokewrap PyPI listing in the Pokewrap section, and add a different background-color to the "Coming Soon" card to set it apart.

Create the Pokewrap project subpage, add breadcrumbs, and begin filling in details.