Refactor css
Make minor changes to CSS to enable it to be more extensible and flexible, and make classes (slightly) more generic so they can be used in more places.

Take a step in the direction of allowing CSS files to be cached longer and, therefore, be more stable. This requires certainty that I won't need to keep adding new CSS, since that won't get served properly when longer cache times are used.