Update site styles & contact page
Adjust contact page to have a smaller, easier-to-use contact form
with a lightly colored border around it. This makes the form look
a bit more cohesive and less stretched, and reduces the distance
between fields on wider desktop views.

Add periods and revise copy throughout site (excluding about page).
Format all text more consistently, include periods because they
help screen readers and the open-endedness of my sentences previously
was a pointless style choice that doesn't matter.

Revise bits of copy throughout the projects page for readability
and tone. It was stilted and incomplete, and now it's less that way!