Adjust content/layout of about page
Reduce content on about page, first pass. Make sections less
wordy and more efficient. Get to the point quicker so people
don't lose interest.

Make skills section a bulleted list for efficiency.

This is the first in a series of changes to update this page,
and is not a final version. Instead, it's the first step
in an iterative journey to have a more compelling about section.