Redesign about page
Use horizontal space and extended class to redesign
the about page. Rearrange the content, reduce
the word count, and be more direct. Create a short
and sweet "Skills" section to promote my core
competencies.

Still a work in progress, but is a step towards
having a cleaner and more focused about page.