Update styles and content sitewide
Adjust styles across various pages to be more
consistent with one another and to better reflect
the different page hierarchies. For example,
about and contact both have similar formats to how
content is displayed, as does the projects page.

Additionally, edit the most recent blog post
for content to ensure it reads well, makes sense,
and hasn't lost touch with its purpose. The new
edits are mild and smooth it over--it was already
quite solid, all things considered.