Adjust blog display styles
Change homepage blog cards to use new classes and
be prettier so as to be more engaging and stand
out from the rest of the page.

Also adjust blog page to work better on desktop
views and condense information to fit more blog
posts in the page since I've been posting
consistently and it's been building up.

The changes from this update are designed to
improve both the browsing and readability of the
blog indices, whether that's a highlight on the
home page or the actual blog index itself.

They're also designed to make future changes
easier, and to future-proof myself as I continue
to publish writing.