Revise home page and footer links
Make hero section more accurate to my current
personal brand and efforts.I'm trying to be both
authentic and open about what I do and what it
means to me.

Update the social links in my footer to include
Instagram, because I might as well, and to be
arranged in a more visually appealing way.