Make entire card clickable
Use pseudo-elements to expand the clickable portion
of each blog preview card so the entire card is
clickable rather than just the title.

Disable focus outline on title and instead focus
the entire card when selected.

This creates a better UX for visitors to be able
to click anywhere on the element that animates,
which makes sense based on their expectations.