Update footer info
Footer is outdated--updated with current social
accounts that I actually check, and changed the
order based on priotity/activity.

Cleaned up icon usage. Note: in the future, icons
may switch to self-hosted SVG and Lucide instead
of a JS library like FontAwesome for performance,
but upon evaluation currently the site is fast
enough it's a negligible change.