Fix stylesheet rendering in 404 page
404 page previously tries to load the relative url
for styles (but does so from whatever incorrect
url is being used) so it often results in an error
and a default html page with no style.

To fix this, use the absolute url to the styles
so that the page will load it correctly regardless
of its relative status in the directory.