Phase 1 Astro rewrite of site
Complete directory overhaul to confirm to Astro
formatting requirements to run and serve content.

Rewriting of styles to encapsulate relevant styles
in their .astro component or layout to reduce
global style bloat.

Rearrange of blog directories to include /posts/
to contain all blog posts for ease of maintenance
and so the content stays maintainable even as the
number of posts increases.

Note: the new file directories also allow for
adding an '/archive/' folder someday to the blog
for archived posts to get them out of the way but
not delete them, thus maintaining link integrity.